CSF425 Deep Learning Project Folder/requirements.txt
ipython
librosa
matplotlib
numpy
optuna
pandas
Pillow
protobuf
pydub
scikit_learn
soundfile
tensorflow
torch
torchaudio
torchvision
opencv-python
setuptools
ffmpeg-python
audioread